Move EigenView to its own header file
