+++ include/ndarray/buildOptions.h.in
@@ -1,0 +1,5 @@
+// Build options that need to be remembered
+// so they can be used when other packages build against ndarray
+
+#cmakedefine NDARRAY_EIGENVIEW
+#cmakedefine NDARRAY_STDPYBIND11EIGEN
